backend/requirements.txt
fastapi==0.116.1
slowapi==0.1.9
uvicorn==0.35.0
pydantic[email]==2.11.7
pydantic-settings==2.10.1
alembic==1.16.4
SQLAlchemy==2.0.42
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
firebase-admin==6.9.0
python-multipart==0.0.20
redis==6.4.0
asyncpg==0.30.0
setuptools==80.9.0
GeoAlchemy2==0.18.0
phonenumberslite==9.0.14
confluent-kafka==2.12.1
orjson==3.11.4
argon2-cffi==25.1.0
